pyproject.toml
[project]
name = "atamai-tamanu-program-forms"
version = "0.1.0"
description = "AI-assisted builder for Tamanu program form XLSX files"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.40.0",
    "anthropic>=0.40.0",
    "baml-py==0.216.0",
    "mcp>=1.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
]

[tool.setuptools.packages.find]
exclude = ["baml_src*", "baml_client*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

